removed moved resources and made resources that used them point to their new shared fhir resource location
