+++ urgent-care/src/main/java/gov/va/api/health/urgentcare/service/controller/capability/MetadataController.java
@@ -9,15 +9,15 @@ import gov.va.api.health.r4.api.datatypes.ContactDetail;
 import gov.va.api.health.r4.api.datatypes.ContactPoint;
 import gov.va.api.health.r4.api.datatypes.ContactPoint.ContactPointSystem;
 import gov.va.api.health.r4.api.elements.Extension;
-import gov.va.api.health.urgentcare.service.api.capability.Capability;
-import gov.va.api.health.urgentcare.service.api.capability.Capability.CapabilityResource;
-import gov.va.api.health.urgentcare.service.api.capability.Capability.ResourceInteraction;
-import gov.va.api.health.urgentcare.service.api.capability.Capability.Rest;
-import gov.va.api.health.urgentcare.service.api.capability.Capability.RestMode;
-import gov.va.api.health.urgentcare.service.api.capability.Capability.SearchParamType;
-import gov.va.api.health.urgentcare.service.api.capability.Capability.Security;
-import gov.va.api.health.urgentcare.service.api.capability.Capability.Software;
-import gov.va.api.health.urgentcare.service.api.capability.Capability.TypeRestfulInteraction;
+import gov.va.api.health.r4.api.resources.Capability;
+import gov.va.api.health.r4.api.resources.Capability.CapabilityResource;
+import gov.va.api.health.r4.api.resources.Capability.ResourceInteraction;
+import gov.va.api.health.r4.api.resources.Capability.Rest;
+import gov.va.api.health.r4.api.resources.Capability.RestMode;
+import gov.va.api.health.r4.api.resources.Capability.SearchParamType;
+import gov.va.api.health.r4.api.resources.Capability.Security;
+import gov.va.api.health.r4.api.resources.Capability.Software;
+import gov.va.api.health.r4.api.resources.Capability.TypeRestfulInteraction;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
